Bug 1865372: Call nsCocoaWindow::DestroyNativeWindow unconditionally in Destroy. r=mac-reviewers,mstange
This seems like the right time to call DestroyNativeWindow since we are
hiding the window and destorying the nsBaseWidget -- there's not much
else that can happen to this window. Making this call here ensures that
the native window is not maintained on the screen waiting for the
destructor to be called during garbage collection.

Differential Revision: https://phabricator.services.mozilla.com/D197214

UltraBlame original commit: 51d87c2129d274294cde7bfa8a52fe1195265a73