Bug 1856590: create a function to return `--remotes` arg and use in `get_branch_nodes` r=ahochheiden
In bug 1866395 I added a function to return the canonical Firefox
remote and use it when determining the base ref. I overlooked
that this also needs to be used for `get_branch_nodes` to avoid
returning a set of branch nodes that doesn't begin from the same
base ref. Move the building of the `--remotes` arg to a separate
function and use it in `get_branch_nodes`.

Differential Revision: https://phabricator.services.mozilla.com/D198032

UltraBlame original commit: 625d99466c9630a42389cdd45b5da207a51bf481