Bug 1865845: stop caching mac attribution data in a separate file r=nalexander
This is mostly removing code and tests related to reading/writing the cache file on macOS and updating of tests. Aside from that, the most notable part is the change to `setAttributionString` to automatically prepend `__MOZCUSTOM__` when writing an attribution code. This is mostly done to make things simpler and cleaner in the majority of the tests, but seeing as `getAttributionString` is also aware of it, it seems like a generally nicer way to do this.

Differential Revision: https://phabricator.services.mozilla.com/D197204

UltraBlame original commit: aa6b62d272a26d974b5afb1b94f65bad67716728