Bug 1871806 - [devtools] Add notice for Enter behavior in Rules view. r=devtools-reviewers,ochameau.
A link points to a blog post explaining why we made the change.
Once the notice is closed by the user, it won't be opened again.
The notice is not localized so it's easier to uplift for now; it will
be localized in a follow up.

Depends on D197897

Differential Revision: https://phabricator.services.mozilla.com/D197924

UltraBlame original commit: 13107fcd50cf2795a397cd8d72ce290a0b34a909