Bug 1871555 - deleteUserInteractionForClearingHistory should appropriately call back on error. r=hsohaney,pbz
This properly forwards rejections to the onDataDeleted() callback...
Unfortunately none of the consumers seem to care, and they likely don't want to
interrupt the ongoing clean up.
This means there may be cases where this data won't be cleared, but they always
existed, and this patch doesn't change that fact.
Not adding a test because it would be extremely synthetic (replacing methods),
thus not particularly robust to future changes.

Differential Revision: https://phabricator.services.mozilla.com/D197152

UltraBlame original commit: 0bd36479d605b1a8748ee37096ff0d54383f0402