Bug 1872398 - Add an exception from a11y_checks for one click on the URL Bar in a mouse-specific test. r=adw
We intentionally turn off a11y_checks for the click updated, because it is send to send a focus on the URL Bar with the mouse, while other ways to focus it are accessible for users of assistive technology and keyboards, thus this test can be excluded from the accessibility tests.

Since this test is now expected to pass a11y_checks, we are also remiving the `fail-if` notation from the test manifest.

Differential Revision: https://phabricator.services.mozilla.com/D197433

UltraBlame original commit: d4319756452b443d8542a89e942aab790df6b48f