Bug 1869043 interleave audio output after AudioMixer mixing r=padenot
This will enable PlayAudio() to be used in a subsequent patch to mix audio
output for CrossGraphReceiver::EnqueueAudio().

An AudioChunk is used to clarify the planar storage and simplify use of
InterleaveAndConvertBuffer().

AudioMixer is also changed to initialize to zero only as many samples in its
buffer as will be used.  Resetting to zero is no longer performed in
FinishMixing() because that is done on first call to Mix().

Differential Revision: https://phabricator.services.mozilla.com/D198220

UltraBlame original commit: 5fed273e4f3d569475b5d41f8e49cec991aa4fbc