Bug 1873794 - Make `HTMLEditor::SplitAncestorStyledInlineElementsAt` stop splitting ancestors of non-splittable elements r=m_kato
The method does not assume that an inline ancestor is not a splittable, but
it may occur in the edge cases.

I tried to make `HTMLEditor` adjust selection range at starting deletion to
stop handling it or select the most distant unmodifiable element, but it
 requires some expectation changes of tentative WPTs. It should be done in
a separate bug instead.

Differential Revision: https://phabricator.services.mozilla.com/D198345

UltraBlame original commit: 23b1e2c5bd986b1d99873297c3cefd98f2b892b9