Bug 1834055 - Drop the hack to be too smart about mouse detection r=win-reviewers,handyman
Trying to see whether the system-reported mouse is a REAL mouse causes problem in certain situation with Wacom driver. The driver makes the system believe it has a pen while it still reports that the input device count is 1. Our heuristic thus thinks that the single device is not a real mouse but something else, which then breaks web compatibility with Twitter.

That hack was added by bug 1493128, but the capability report behavior change is already reverted with bug 1813979. Keeping this hack thus has no point as of now. This patch follows the comment: "We shouldn't assume we know better."

Differential Revision: https://phabricator.services.mozilla.com/D197813

UltraBlame original commit: f69fdcb8418b22a6be5afbf9019183539cbdb4de