Bug 1875194 - Fix vertical scrollbar size returned by GetDesiredScrollbarSizes(). r=dholbert
When the scrollbar is on the right hand side, its size should be set in `right`
instead of in `left`.

This bug is not observable because all the callers are accessing the result
directly via `LeftRight()`, or they are transforming the result into a
`LogicalMargin` and access it via `IStartEnd()` or `BStartEnd()`. However, it is
still worth fixing for the future.

Differential Revision: https://phabricator.services.mozilla.com/D198907

UltraBlame original commit: d5a5014e5e3e1804067092872415d6c2fe746beb