Bug 1863543 part 2 - Optimize CacheIR for missing properties on DOM proxies. r=iain,peterv
We were optimizing this when we found the property on the proto chain, but
if the property was missing we were emitting a slow VM call. We can optimize
this in a similar way for override-builtins proxies such as `HTMLDocument`.

Differential Revision: https://phabricator.services.mozilla.com/D192956

UltraBlame original commit: 2d3e18108992ae4dba1711ac01d915360ffc9ab6