Bug 1874727 - Make `NativeKey::ComputeInputtingStringWithKeyboardLayout()` compute shortcut candidates when both `Ctrl` and `Alt` are pressed r=m_kato
It does not compute shortcut candidates when both `Ctrl` and `Alt` are pressed.
This was valid before bug 900750, but when the method is called, now, `Ctrl`
and `Alt` are unset if `AltGr` is pressed and that introduces a character or
a dead character.
https://searchfox.org/mozilla-central/rev/8b6cf2a8352979c781dfd90c46a9f4f7135fabea/widget/windows/KeyboardLayout.cpp#1663-1666,1683,1688-1689

Therefore, it should be safe to compute shortcut candidates when `Ctrl` and
`Alt` are pressed even though such shortcut keys shouldn't be used from i18n
point of view.

Differential Revision: https://phabricator.services.mozilla.com/D198659

UltraBlame original commit: 938e67ed7818cdb5cabd386af7b01ed348094492