Bug 1824671 - patch 5 - Update WPT metadata for tests affected by the change in bidi implementation. r=platform-i18n-reviewers,dminor
The failure in bidi-011.xht occurs because unicode-bidi (unlike ICU4C) closely follows the
implementation notes in UAX#9 regarding treatment of the bidi embedding/override controls,
if the characters are not actually removed from the buffer during processing. The (non-normative)
algorithm given in the spec results in embedding initiators taking on the level of the context,
whereas the legacy ICU4C behavior gives them the level of the embedded (or overridden) text.
This difference affects bidi continuation creation when the directional controls occur exactly
at inline element boundaries, and is observable in this test.

We could fix this by post-processing the unicode-bidi results to conform to what ICU4C did,
but I don't think it's worth it for this extremely niche case (which is not mandated by any
spec, AFAICT; and it's unclear whether any particular behavior is better or worse than any
other). Moreover, applying a tweak to fix this will also revert the newly-passing examples
in css-text/white-space/white-space-collapse-002.html, so we'd lose as much as we'd gain.

The new failures in first-letter-punctuation-126 and -127 are because unlike ICU4C, unicode-bidi
(correctly, per spec) assigns level 2 to characters with class AN where the surrounding context
is level 0. These two examples use Arabic numeric punctuation characters that have class AN in
conjunction with a Latin letter (class L), and so we get a bidi level boundary between the letter
and the punctuation, which interferes with our first-letter implementation.

This is such an obscure edge case, extremely unlikely to ever occur in real content, that I'm happy
just marking it as an expected failure for now. Our entire first-letter implementation has much
more serious issues than this.

Differential Revision: https://phabricator.services.mozilla.com/D197892

UltraBlame original commit: e16c7ccc64e5ef2204da2ea255efca70ef8995b1