Bug 1873776 - wasm: Fix Ion bug with throw_ref inside a try. r=jseward
A throw_ref in Ion inside of a `try` or `try_table` would emit a call
to Instance::throwException to set the pending exception and
trigger the unwinding. It does this instead of using the Ion
FunctionCompiler::throwFrom helper, because we don't have the
tag of the exnref laying around, and getting it is difficult.

However, Ion MIR generation has an invariant that function calls
that throw generate a control flow graph edge to their nearest
enclosing try. This is setup today for wasm calls, but not for
wasm instance method calls. This commit updates wasm instance
method calls to generate the right MIR graph.

Differential Revision: https://phabricator.services.mozilla.com/D198093

UltraBlame original commit: 900e842c161345646e05c7d557b370f30b7d80ea