Bug 1875955 - Support the "optional" search engine configuration flag. r=Standard8
At the moment, desktop does not have an optional search engine flag to add an
optional engine from the list of engines. This is a feature only on Android.
We want to skip the optional flag in the new search config when the user is on
firefox desktop.

Differential Revision: https://phabricator.services.mozilla.com/D199296

UltraBlame original commit: 98fcc4fc24ca8de3ad88764a67671c702facf772