Bug 1842773 - Part 23: Add TypedArrayLengthZeroOnOutOfBounds. r=sfink,jandem
Part 21 changed the `TypedArrayLength` to throw for out-of-bounds typed arrays,
but in a handful of cases we actually need to read the typed array length
without throwing an error for out-of-bounds.

Add a separate intrinsic for these cases and update the relevant self-hosted
functions to call `TypedArrayLengthZeroOnOutOfBounds`.

Differential Revision: https://phabricator.services.mozilla.com/D183340

UltraBlame original commit: 4dd54c5f0b0f25f8aef760da4611179c4893d64e