Bug 728807 - Adjust containing block origin for fixed positioned elements. r=layout-reviewers,emilio,dholbert
The containing block rect which the fixed position elements are relative to
should exclude the scrollbar or scrollbar-gutter area. Current, we always use
the origin at (0,0), which makes fixed position elements draw behind the gutter
area on the left edge of the viewport.

This patch adjusts the fixed pos containing block's origin when there is a
scrollbar or scrollbar-gutter on the left edge of the viewport. In other words,
the containing block's origin becomes the same as `nsCanvasFrame`'s origin.

The "-001" test uses `scrollbar-gutter: stable`, and we already pass it without
this patch. The "-002" variant tests `scrollbar-gutter: stable both-edges`,
which requires this patch to pass.

Differential Revision: https://phabricator.services.mozilla.com/D199687

UltraBlame original commit: 34fcdc11eb39d6b7f124a12120bc8f27f2bf8e60