Bug 1873255 - Update 6 mochitests that click on an addon-name-container instead of its containing link to pass a11y-checks. r=rpl
There are following tests in `toolkit/mozapps/extensions/test/browser/` that fail a11y_checks when they send a click on the clickable but not accessible container `.addon-name-container` in:
1. [browser_globalwarnings.js#11](https://searchfox.org/mozilla-central/rev/9c509b8feb28c1e76ad41e65bf9fd87ef672b00f/toolkit/mozapps/extensions/test/browser/browser_globalwarnings.js#11),
1. [browser_html_detail_view.js#346,405](https://searchfox.org/mozilla-central/rev/9c509b8feb28c1e76ad41e65bf9fd87ef672b00f/toolkit/mozapps/extensions/test/browser/browser_html_detail_view.js#346,405),
1. [browser_html_updates.js#31](https://searchfox.org/mozilla-central/rev/9c509b8feb28c1e76ad41e65bf9fd87ef672b00f/toolkit/mozapps/extensions/test/browser/browser_html_updates.js#31)
1. [browser_gmpProvider.js#52](https://searchfox.org/mozilla-central/rev/9c509b8feb28c1e76ad41e65bf9fd87ef672b00f/toolkit/mozapps/extensions/test/browser/browser_gmpProvider.js#52)
1. [switchToDetailView() in head.js#1696](https://searchfox.org/mozilla-central/rev/9c509b8feb28c1e76ad41e65bf9fd87ef672b00f/toolkit/mozapps/extensions/test/browser/head.js#1696) in `browser_html_pending_updates.js`
1. `testCardMoreOptionsButton()` in `browser_menu_button_accessibility.js`
- while there is a fully accessible link element that is included in each card and thus should be an expected target for the Details View activation to test behavior for all users is functioning.

The test `browser_html_detail_view.html` would still click on the container element to ensure this user-facing behavior is also covered with mochitests, while on-link clicks are available.

We could update this test to query the link element and send a click to it to simulate and confirm the card is accessible for users of keyboards and assistive technology as well as mouse/touch users, and to allow the a11y-checks to pass. In this case, we need to update the `fail-if` notations of the test manifest accordingly.

Differential Revision: https://phabricator.services.mozilla.com/D197811

UltraBlame original commit: b0a09cf1582ee4f7fcabed367057fa534ea175c2